// src/app/interfaces/curso.ts
export interface Curso {
    idarticulos: number;
    curso: string
    nocertificado: string
    fecha: string
    fecha_registro: string
    estatus: string
    codigo: string
    descripcion: string
    precio_mayoreo: number
    id_pay: string
    status_pay: string
    nombre_pay: string
    email_pay: string
    total_pay: number
    method_pay: string
}